Modified bad_actor.py so it produces an astropy table
The information in the table is identical to that in bad_actor_all.txt,
but a table output is more convenient for other analyis such as
comparisons with the bad actor prediction tool.  It is not entirely
obvious that we need to keep bad_actor_all.txt the pure ascii file,
but the ascii file is created on the fly, while the table is created
at the end of the routine.