Allow user control of the lookback time
A new input was added to subtract_persist.do_dataset which allows the
user to vary the lookback time for observations that could cause
persistrence.  The default remains 16 hours.  A -lookback switch has
been added both to run_persist.py and subtract_persist.py so that the
lookback time can be changed from the command line.  The lookback time
has also been added to the history log, and hence to the html file
produced by run_persist.py