Updates run_persist so that it can be run with multiple processors
An additional switch -np has been added that controls the number of
datasets being processed simultaneously.  Changes in other routines are
minor and mainly intended to reduce the information sent to the terminal

These routines still need to be tested, particularly in situations where
the observations.sum file is large.  Nothing has been done yet to assure
that there are not collisions associated with this.