Update to per_list.py so that the most time intensive part can be run with multiple processors
Split out the the portion of make_ordered list that reads the keywords
into a new function get_info and used the Pool mechanism to be able to
run this in parallel.  A new switch has been added -np to the input file
which sets the number of simulataneous processes to use.