Updated all of the actually run_persist code to python3. Fixed various portions of the code to use astropy tables.
There are still areas wwhere tables are not use, and the code works
only because the the astropy table format is not very different from a
flat file.  An important example of this is updating the obsum file.

This code does not yet include a planned fix to speed up per_list