Fix to per_list.py to handle duplicates properly when using join
When running per_list on large sets of data (e.g the entire archive), per_list
was finding large numbers of duplicates. This happened in part because left
joins do not necessarily preserve the order of lines in a table.  The join
had been introduced to speed up s matching operation.  See #29